requirements.txt
blinker==1.4
Click==7.0
dnspython==1.16.0
Flask==1.1.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-WTF==0.14.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pymodm==0.4.2
pymongo==3.10.1
python-dotenv==0.11.0
Werkzeug==0.16.1
WTForms==2.2.1
